Keep and Change Auto Run
Versions of my Monty Hall game but the initial door choice is randomised in both and one program always keeps the original door and the other always changes to the other door. This was used in a for loop to automate data collection for statistical analysis of the Monty Hall Problem.